admin-panel/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
django-filter>=25.2,<26
dj-database-url>=3.1,<4
psycopg[binary]>=3.2,<4
Pillow>=12,<13
openpyxl>=3.1,<4
